Rewrite streaming iterators, use an internal circular buffer
BufRead has a fill_buf/consume logic that seemed what we want, but does
only read data by chunks. We parsing data, we can be before end of
buffer, which creates a problem: next read is Incomplete, but if we read
data, the end of the buffer will be lost (and we can't store it inside
the iterator).

To solve this, use the circular crate, so we can do the same, and simply
shift data when the buffer is partially consumed.