Use git-cliff to generate CHANGELOG (with some manual editions)
Use git-cliff for automatic generation of the Changelog, in the 'Keep a
Changelog' format.
However, we do not use it in fully automatic mode: previous changes are not all
in the correct format and this would loose some content (but do we care?).
This may be considered again in the future to switch to fully automatic.